スキルの調整、仮のタイトルとロゴシーンを追加、キャラクターのマテリアル設定

+++ CreatureBattle/Assets/Scripts/UI/HpBar.cs
@@ -17,6 +17,6 @@ public class HpBar : MonoBehaviour {
     }
 
     void LateUpdate(){
-        test.text = _actor.GetStatus().GetHP().ToString();
+        test.text = ((int)_actor.GetStatus().GetHP()).ToString();
     }
 }
